+++ var/spack/repos/builtin/packages/octave-optim/package.py
@@ -23,10 +23,9 @@
 # Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307 USA
 ##############################################################################
 from spack import *
-import os
 
 
-class OctaveOptim(Package):
+class OctaveOptim(OctavePackage):
     """Non-linear optimization toolkit for Octave."""
 
     homepage = "https://octave.sourceforge.io/optim/"
@@ -35,16 +34,4 @@ class OctaveOptim(Package):
     version('1.5.2', 'd3d77982869ea7c1807b13b24e044d44')
 
     depends_on('octave-struct@1.0.12:')
-
     extends('octave@3.6.0:')
-
-    def install(self, spec, prefix):
-        os.environ.pop('CC', '')
-        os.environ.pop('CXX', '')
-        os.environ.pop('FC', '')
-        octave('--quiet',
-               '--norc',
-               '--built-in-docstrings-file=/dev/null',
-               '--texi-macros-file=/dev/null',
-               '--eval', 'pkg prefix %s; pkg install %s' %
-               (prefix, self.stage.archive_file))
