+++ var/spack/repos/builtin/packages/octave-splines/package.py
@@ -25,20 +25,11 @@
 from spack import *
 
 
-class OctaveSplines(Package):
+class OctaveSplines(OctavePackage):
     """Additional spline functions."""
 
     homepage = "http://octave.sourceforge.net/splines/index.html"
     url      = "http://downloads.sourceforge.net/octave/splines-1.3.1.tar.gz"
 
     version('1.3.1', 'f9665d780c37aa6a6e17d1f424c49bdeedb89d1192319a4e39c08784122d18f9')
-
     extends('octave@3.6.0:')
-
-    def install(self, spec, prefix):
-        octave('--quiet',
-               '--norc',
-               '--built-in-docstrings-file=/dev/null',
-               '--texi-macros-file=/dev/null',
-               '--eval', 'pkg prefix %s; pkg install %s' %
-               (prefix, self.stage.archive_file))
